python/requirements.txt
# mandatory dependencies
pre-commit~=4.2.0
python-dotenv~=1.1.0

# doc reader finetuning
datasets~=3.3.1
transformers~=4.50.0
huggingface-hub~=0.29.0
torch~=2.6.0
accelerate~=1.6.0
